More brightness calculation changes
