+++ src/main/kotlin/color/BrightnessCalculator.kt
@@ -1,19 +1,27 @@
-package color.brightness
+package color
 
 import config.ConfigSingleton
 import sound.frequencyBins.FrequencyBinList
+import sound.frequencyBins.GreatestMagnitudeCalculator
 import sound.frequencyBins.filters.BandPassFilter
 import sound.frequencyBins.filters.Crossover
 
-class GreatestMagnitudeCalculator(
+class BrightnessCalculator(
     private val bandPassFilter: BandPassFilter = BandPassFilter(),
     private val lowCrossover: Crossover = ConfigSingleton.lowCrossover,
     private val highCrossover: Crossover = ConfigSingleton.highCrossover,
+    private val greatestMagnitudeCalculator: GreatestMagnitudeCalculator = GreatestMagnitudeCalculator(),
 ) {
 
     fun calculate(frequencyBins: FrequencyBinList): Float? {
         val filteredBins = getFilteredBins(frequencyBins)
-        return filteredBins.maxOfOrNull { it.magnitude }
+        val magnitude = greatestMagnitudeCalculator.calculate(filteredBins) ?: return null
+
+        return if (magnitude < 1F) {
+            magnitude
+        } else {
+            1F
+        }
     }
 
     private fun getFilteredBins(frequencyBins: FrequencyBinList): FrequencyBinList {
@@ -25,3 +33,4 @@ class GreatestMagnitudeCalculator(
     }
 
 }
+
