+++ src/main/kotlin/sound/frequencyBins/GreatestMagnitudeCalculator.kt
@@ -1,0 +1,9 @@
+package sound.frequencyBins
+
+class GreatestMagnitudeCalculator {
+
+    fun calculate(frequencyBins: FrequencyBinList): Float? {
+        return frequencyBins.maxOfOrNull { it.magnitude }
+    }
+
+}
