+++ src/test/kotlin/color/ColorFactoryTests.kt
@@ -1,7 +1,5 @@
 package color
 
-import color.brightness.BrightnessCalculator
-import color.hue.HueCalculator
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
