+++ src/test/kotlin/color/HueCalculatorIntegrationTests.kt
@@ -1,4 +1,4 @@
-package color.hue
+package color
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
